+++ src/modules/constants.mjs
@@ -24,38 +24,38 @@ const tah = {
    * Action types
    */
   actions: {
-    ability: "tokenActionHud.black-flag.ability",
-    check: "tokenActionHud.black-flag.check",
-    counter: "tokenActionHud.black-flag.counter",
-    effect: "tokenActionHud.black-flag.effect",
-    exhaustion: "tokenActionHud.black-flag.exhaustion",
-    feature: "tokenActionHud.black-flag.feat",
-    save: "tokenActionHud.black-flag.save",
-    skill: "tokenActionHud.black-flag.skill",
-    spell: "tokenActionHud.black-flag.spell",
-    item: 'tokenActionHud.black-flag.item',
-    utility: 'tokenActionHud.utility',
+    ability: "tokenActionHud.black-flag.actions.ability",
+    check: "tokenActionHud.black-flag.actions.check",
+    feature: "tokenActionHud.black-flag.actions.feature",
+    save: "tokenActionHud.black-flag.actions.save",
+    skill: "tokenActionHud.black-flag.actions.skill",
+    spell: "tokenActionHud.black-flag.actions.spell",
+    item: 'tokenActionHud.black-flag.actions.item',
+    container: 'tokenActionHud.black-flag.actions.container',
+    utility: 'tokenActionHud.utility'
   },
 
   /**
    * Groups
    */
   groups: {
-    abilities: {id: 'abilities', name: 'tokenActionHud.black-flag.abilities', type: 'system'},
-    skill: {id: 'skill', name: 'tokenActionHud.black-flag.skill', type: 'system'},
-    vehicle: {id: 'vehicle', name: 'tokenActionHud.black-flag.vehicle', type: 'system'},
+    checks: {id: 'checks', name: 'tokenActionHud.black-flag.checks', type: 'system'},
+    saves: {id: 'saves', name: 'tokenActionHud.black-flag.saves', type: 'system'},
+    skills: {id: 'skills', name: 'tokenActionHud.black-flag.skills', type: 'system'},
 
-    actions: {id: 'actions', name: 'tokenActionHud.black-flag.actions', type: 'system'},
+    vehicle: {id: 'vehicle', name: 'BF.VEHICLE.Label[other]', type: 'system'},
 
-    inventory: {id: 'inventory', name: 'tokenActionHud.black-flag.inventory', type: 'system'},
-    containers: {id: 'containers', name: 'tokenActionHud.black-flag.containers', type: 'system'},
-    ammunition: {id: 'ammunition', name: 'tokenActionHud.black-flag.ammunition', type: 'system'},
-    armor: {id: 'armor', name: 'tokenActionHud.black-flag.armor', type: 'system'},
-    consumables: {id: 'consumables', name: 'tokenActionHud.black-flag.consumables', type: 'system'},
-    gear: {id: 'gear', name: 'tokenActionHud.black-flag.gear', type: 'system'},
-    sundries: {id: 'sundries', name: 'tokenActionHud.black-flag.sundries', type: 'system'},
-    tools: {id: 'tools', name: 'tokenActionHud.black-flag.tools', type: 'system'},
-    weapons: {id: 'weapons', name: 'tokenActionHud.wfrp4e.weapons', type: 'system'},
+    actions: {id: 'actions', name: 'BF.ACTIVATION.Type.Action[other]', type: 'system'},
+
+    inventory: {id: 'inventory', name: 'BF.Sheet.Tab.Inventory', type: 'system'},
+    containers: {id: 'containers', name: 'BF.Item.Type.Container[other]', type: 'system'},
+    ammunition: {id: 'ammunition', name: 'BF.Item.Type.Ammunition[other]', type: 'system'},
+    armor: {id: 'armor', name: 'BF.Item.Type.Armor[other]', type: 'system'},
+    consumables: {id: 'consumables', name: 'BF.Item.Type.Consumable[other]', type: 'system'},
+    gear: {id: 'gear', name: 'BF.Item.Type.Gear[other]', type: 'system'},
+    sundries: {id: 'sundries', name: 'BF.Item.Type.Sundry[other]', type: 'system'},
+    tools: {id: 'tools', name: 'BF.Item.Type.Tool[other]', type: 'system'},
+    weapons: {id: 'weapons', name: 'BF.Item.Type.Weapon[other]', type: 'system'},
 
     utility: {id: 'utility', name: 'tokenActionHud.utility', type: 'system'},
     combat: {id: 'combat', name: 'tokenActionHud.black-flag.combat', type: 'system'},