+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
+## v0.0.2
+- Module currently offers:
+  - Ability Checks
+  - Ability Saves
+  - Skill Checks
+  - Inventory tree with containers
+    - Posting to Chat / Rendering Sheet
+  - Utility > Combat
+
 ## v0.0.1
 - Module currently offers: 
-  - Utility > Combat
   - Displays Inventory tree with containers
+  - Utility > Combat